requirements.txt
asgiref==3.7.2
Django==5.0.1
easy-thumbnails==2.8.5
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.9.0
